+++ ReleaseNotes.md
@@ -4,6 +4,12 @@
 
 ## Release 1.1.1
 - No value checks for compositionId for uuid anymore
+- setting new ressources in helm chart
+- setting new timeouts and retries in helm chart
+- updated dependencies
+- Updated ospo-resources for adding additional notes and disclaimer
+- New realm compatibility
+
 
 ## Release 1.1.0 (20.02.2025)
 - First official GitHub-Release